#487: moved args and transformer
