+++ src/backend/src/prisma-query-args/description-bullets.query-args.ts
@@ -1,0 +1,5 @@
+import { Prisma } from '@prisma/client';
+
+export const descBulletArgs = Prisma.validator<Prisma.Description_BulletArgs>()({
+  include: { userChecked: true }
+});
