+++ src/backend/src/routes/description-bullets.routes.ts
@@ -1,6 +1,6 @@
 import express from 'express';
 import { body } from 'express-validator';
-import { checkDescriptionBullet } from '../controllers/description-bullets.controllers';
+import DescriptionBulletController from '../controllers/description-bullets.controllers';
 import { validateInputs } from '../utils/utils';
 
 const descriptionBulletsRouter = express.Router();
@@ -10,7 +10,7 @@ descriptionBulletsRouter.post(
   body('userId').isInt({ min: 0 }).not().isString(),
   body('descriptionId').isInt({ min: 0 }).not().isString(),
   validateInputs,
-  checkDescriptionBullet
+  DescriptionBulletController.checkDescriptionBullet
 );
 
 export default descriptionBulletsRouter;
