+++ src/backend/src/utils/description-bullets.utils.ts
@@ -1,19 +1,5 @@
 import prisma from '../prisma/prisma';
-import { Prisma, Role } from '@prisma/client';
-
-export const descBulletArgs = Prisma.validator<Prisma.Description_BulletArgs>()({
-  include: { userChecked: true }
-});
-
-export const descBulletTransformer = (descBullet: Prisma.Description_BulletGetPayload<typeof descBulletArgs>) => {
-  return {
-    id: descBullet.descriptionId,
-    detail: descBullet.detail,
-    dateAdded: descBullet.dateAdded,
-    dateDeleted: descBullet.dateDeleted ?? undefined,
-    userChecked: descBullet.userChecked ?? undefined
-  };
-};
+import { Role } from '@prisma/client';
 
 export const hasBulletCheckingPermissions = async (userId: number, descriptionId: number) => {
   const user = await prisma.user.findUnique({ where: { userId } });
